+++ debian/control
@@ -1,4 +1,4 @@
-Source: power-toggle-switch
+Source: msu-manager
 Section: python
 Priority: optional
 Maintainer: Markus Zarbock <foss@starwit.de>
@@ -6,11 +6,11 @@ Build-Depends: debhelper (>= 11), dh-python, python3-all (>= 3.10), python3-setu
 Standards-Version: 4.1.3
 Homepage: https://starwit.de
 
-Package: power-toggle-switch
+Package: msu-manager
 Architecture: all
 Depends: ${misc:Depends}, ${python3:Depends}, python3-all (>= 3.10), python3-pip,
         python3-venv
 XB-Python-Version: ${python:Versions}
-Description: Power toggle service
- This is a service, that can turn power for a server off. It listens on a
- network socket for incoming requests. 
+Description: MSU Manager
+ This is a service, that manages Mobilytix Sensoring Units. It controls power and monitors microcontroller,
+ that manages power and thermal management.
