+++ debian/postinst
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 
-SERVICE_NAME=power-toggle-switch
+SERVICE_NAME=msu-manager
 
 # Install Python packages that aren't available via APT
 if [ "$1" = "configure" ]; then
