--- debian/power-toggle-switch.service
@@ -1,14 +1,0 @@
-[Unit]
-Description=Power Toggle Service
-After=network.target
-
-[Service]
-Type=simple
-User=powertoggle
-WorkingDirectory=/opt/power-toggle-switch
-ExecStart=/usr/local/bin/power-toggle-switch
-Restart=always
-RestartSec=5
-
-[Install]
-WantedBy=multi-user.target
