+++ main.py
@@ -1,4 +1,4 @@
-from power_toggle_switch import run_stage
+from msu_manager import run_stage
 
 if __name__ == '__main__':
     run_stage()
