+++ msu_manager/__init__.py
@@ -1,0 +1,3 @@
+from .msu_manager import run_stage
+
+__all__ = ['run_stage']
