--- power_toggle_switch/__init__.py
@@ -1,3 +1,0 @@
-from .power_toggle_switch import run_stage
-
-__all__ = ['run_stage']
